Bring clock module back but the code doesn't work due to the 64bit integer divide operation. Fix it in later version.

+++ include/types.h
@@ -33,6 +33,6 @@ typedef uint32_t phys_size_t;
 typedef int ptrdiff_t;
 
 /* Micro second type definition */
-typedef uint32_t useconds_t;
+typedef uint64_t useconds_t;
 
 #endif	/* __TYPES_H__ */
